modified problem1.js
Removed a 'validateInventory'function . Added a import using const 'validateInventory'. created a function 'getCarById' for finding the car form inventorty using ID. The if condition checks car id is a number means it use a for loop to identify the given input stores a value in inventory[i].otherwise it pass a invalid input:string.