modified problem6.js added.
Removed a 'validateInventory'function . Added a import using const 'validateInventory'and also removed unused function of getallcars. created a function 'getOldCars' to identify the car year before year of 2000 in inventory .forloop using 'inventory.length' to identify the count of cars and below year of 2000.If condition will apply for the print a format of array. changed all variable names in camelCase method.